requirements.txt
aiohttp~=3.7.4.post0
python-dateutil~=2.8.1
pytz~=2021.1
requests~=2.25.1
